+++ src/renderer/CategoryPage.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import Box from '@mui/material/Box';
-import { AppState, RendererVideo } from 'main/types';
+import { AppState, RecStatus, RendererVideo } from 'main/types';
 import {
   Button,
   List,
@@ -9,7 +9,7 @@ import {
   Typography,
 } from '@mui/material';
 import { scrollBarSx } from 'main/constants';
-import { MutableRefObject, useState } from 'react';
+import { MutableRefObject } from 'react';
 import { VideoPlayer } from './VideoPlayer';
 import { VideoCategory } from '../types/VideoCategory';
 import SearchBar from './SearchBar';
@@ -32,6 +32,7 @@ interface IProps {
   setAppState: React.Dispatch<React.SetStateAction<AppState>>;
   persistentProgress: MutableRefObject<number>;
   playerHeight: MutableRefObject<number>;
+  recorderStatus: RecStatus;
 }
 
 /**
@@ -46,6 +47,7 @@ const CategoryPage = (props: IProps) => {
     setAppState,
     persistentProgress,
     playerHeight,
+    recorderStatus,
   } = props;
   const { numVideosDisplayed, videoFilterQuery } = appState;
   const [config, setConfig] = useSettings();
@@ -138,6 +140,7 @@ const CategoryPage = (props: IProps) => {
             setAppState={setAppState}
             selected={selected}
             persistentProgress={persistentProgress}
+            recorderStatus={recorderStatus}
           />
         </ListItemButton>
       </ListItem>